adds note for skipped tests, updates maxPyrHt test
These three maxPyrHt tests were failing because there's a difference
between Eero's matlab implementation and this python one. Based on the
results of actually calling `ppt.Lpyr((10, 10), filter1='binom4')`
(and similarly for the others), I think python's maxPyrHt is doing the
right thing, so I changed the values to reflect that.