resolves segfault issue
with this commit, we fix that occasional segfault. it was happening when
both the image and the filter were 1 dimensional but they're dimensions
were misaligned. there was already code in Gpyr, Lpyr, and Wpyr to
handle these cases, but the one in Lpyr was not working correctly. I
tried to implement a general fix in corrDn and upConv (the functions
where the segfault actually occured), but this didn't work for some
reason. instead, the best way to handle this is to reshape the filters
during pyramid construction.

this changes the code in Gpyr and Lpyr to be a litle more robust,
restores the tests in unitTests.py that would catch the error for
Lpyr (separate tests already exist that would catch it for Gpyr and
Wpyr), adds a new test for Lpyr to explicitly check the filter shape,
and removes the ineffective code from corrDn and upConv