Updates showIm
this gets the bulk of work in showIm done, by making sure the axes are
the proper size (an integer multiple of the size of the image). I think
this works, in an OS-independent method, but it needs serious
testing. several more features also need to be implemented with showIm