several updates to display_tools
three major changes here:

1. animshow: time now on first dimension, make suse of make_figure and
new function find_zooms

2. imshow can now handle images of different sizes. for each smaller
image, there must be an integer divisor so that we can scale them up to
match the largest image. we also now correctly display the zoom for each
image

3. if imshow's title argument is None, we no longer add the extra space
on top of each axis. we do this with a new argument to make_figure,
vert_pct, which gets set to 1 to not add any space