replaces SteerablePyramidComplex
this commit gets rid of the SteerablePyramidComplex class, because it's
just a special case of the SteerablePyramidFreq class. Therefore, the
same thing can be achieved by calling that class with
`is_complex=True`. Several other small changes to account for
that (renaming the tests, using pyr.is_complex in display_tools.pyrshow,
etc)