pyrshow can handle 1d pyramids
with this commit, pyrshow can handle 1d pyramids. for now, it's pretty
simple, just plotting them as scales by orientations, not allowing the
user to over-ride this behavior, but will pass kwargs on.

find_zooms should now also work for images of any number of
dimensions, though we don't use this