adds better test for supported type
change _convert_signal_to_list so that it checks whether it's a list or
an array, and that's it. previously, it would check the dimensionality
of the array (if it was a single array, but nothing if it was a list)
and would allow any non-arrays non-lists to pass on through

specifically, this means we no longer support passing tensors. reasoning
being that the behavior was potentially unexpected, and we're working on
plenoptic.imshow, which will explicitly support plotting them